+++ app/src/androidTestKeyboards/kotlin/be/scri/ui/screens/settings/SettingsScreenInstallKeyboardButtonTest.kt
@@ -29,6 +29,7 @@ class SettingsScreenInstallKeyboardButtonTest {
         every { mockViewModel.vibrateOnKeypress } returns MutableStateFlow(false)
         every { mockViewModel.popupOnKeypress } returns MutableStateFlow(false)
         every { mockViewModel.isUserDarkMode } returns MutableStateFlow(false)
+        every { mockViewModel.isIncreaseTextSize } returns MutableStateFlow(false)
         return mockViewModel
     }
 
@@ -43,6 +44,7 @@ class SettingsScreenInstallKeyboardButtonTest {
                 SettingsScreen(
                     onDarkModeChange = {},
                     onLanguageSettingsClick = {},
+                    onIncreaseTextSizeChange = {},
                     context = context,
                     viewModel = mockViewModel,
                 )
@@ -68,6 +70,7 @@ class SettingsScreenInstallKeyboardButtonTest {
                 SettingsScreen(
                     onDarkModeChange = {},
                     onLanguageSettingsClick = {},
+                    onIncreaseTextSizeChange = {},
                     context = context,
                     viewModel = mockViewModel,
                 )
@@ -90,6 +93,7 @@ class SettingsScreenInstallKeyboardButtonTest {
                 SettingsScreen(
                     onDarkModeChange = {},
                     onLanguageSettingsClick = {},
+                    onIncreaseTextSizeChange = {},
                     context = context,
                     viewModel = mockViewModel,
                 )
@@ -113,6 +117,7 @@ class SettingsScreenInstallKeyboardButtonTest {
                 SettingsScreen(
                     onDarkModeChange = {},
                     onLanguageSettingsClick = {},
+                    onIncreaseTextSizeChange = {},
                     context = context,
                     viewModel = mockViewModel,
                 )
@@ -137,6 +142,7 @@ class SettingsScreenInstallKeyboardButtonTest {
                 SettingsScreen(
                     onDarkModeChange = {},
                     onLanguageSettingsClick = {},
+                    onIncreaseTextSizeChange = {},
                     context = context,
                     viewModel = mockViewModel,
                 )
@@ -160,6 +166,7 @@ class SettingsScreenInstallKeyboardButtonTest {
                 SettingsScreen(
                     onDarkModeChange = {},
                     onLanguageSettingsClick = {},
+                    onIncreaseTextSizeChange = {},
                     context = context,
                     viewModel = mockViewModel,
                 )
@@ -186,6 +193,7 @@ class SettingsScreenInstallKeyboardButtonTest {
                 SettingsScreen(
                     onDarkModeChange = {},
                     onLanguageSettingsClick = {},
+                    onIncreaseTextSizeChange = {},
                     context = context,
                     viewModel = mockViewModel,
                 )
@@ -210,6 +218,7 @@ class SettingsScreenInstallKeyboardButtonTest {
                 SettingsScreen(
                     onDarkModeChange = {},
                     onLanguageSettingsClick = {},
+                    onIncreaseTextSizeChange = {},
                     context = context,
                     viewModel = mockViewModel,
                 )