+++ app/src/keyboards/java/be/scri/activities/MainActivity.kt
@@ -61,6 +61,14 @@ class MainActivity : ComponentActivity() {
                             .getUserDarkModePreference(context) == AppCompatDelegate.MODE_NIGHT_YES,
                     )
                 }
+
+            val isIncreaseTextSize =
+                remember {
+                    mutableStateOf(
+                        PreferencesHelper.getIncreaseTextSizePreference(context),
+                    )
+                }
+
             val pagerState =
                 rememberPagerState {
                     screens.size
@@ -87,12 +95,18 @@ class MainActivity : ComponentActivity() {
 
             ScribeTheme(
                 useDarkTheme = isDarkMode.value,
+                isIncreaseTextSize = isIncreaseTextSize.value,
             ) {
                 ScribeApp(
                     pagerState = pagerState,
                     isDarkTheme = isDarkMode.value,
+                    isIncreaseTextSize = isIncreaseTextSize.value,
                     onDarkModeChange = { darkMode ->
                         updateTheme(darkMode)
+                    },
+                    onIncreaseTextSizeChange = { increaseTextSize ->
+                        PreferencesHelper.setIncreaseTextSizePreference(context, increaseTextSize)
+                        isIncreaseTextSize.value = increaseTextSize
                     },
                     resetHints = {
                         isHintChangedMap[0] = true