+++ app/src/main/java/be/scri/ui/screens/about/AboutScreen.kt
@@ -17,6 +17,9 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import be.scri.R
 import be.scri.helpers.PreferencesHelper
+import be.scri.helpers.AppFlavor
+import be.scri.helpers.FlavorProvider
+import be.scri.helpers.ui.HintUtils
 import be.scri.ui.common.ScribeBaseScreen
 import be.scri.ui.common.components.ItemCardContainerWithTitle
 import be.scri.ui.screens.about.AboutUtil.getCommunityList
@@ -36,6 +39,7 @@ fun AboutScreen(
     context: Context,
     modifier: Modifier = Modifier,
 ) {
+    val isConjugateApp = FlavorProvider.get() == AppFlavor.CONJUGATE
     val scrollState = rememberScrollState()
 
     val communityList =
@@ -43,8 +47,9 @@ fun AboutScreen(
             onWikimediaAndScribeClick = {
                 onWikiClick()
             },
-            onShareScribeClick = { AboutUtil.onShareScribeClick(context) },
+            onShareScribeClick = { AboutUtil.onShareScribeClick(context, isConjugateApp) },
             context = context,
+            isConjugateApp = isConjugateApp,
         )
 
     val feedbackAndSupportList =
@@ -56,6 +61,7 @@ fun AboutScreen(
                 resetHints()
             },
             context = context,
+            isConjugateApp = isConjugateApp,
         )
 
     val legalItemsList =